requirements.txt
alembic==1.1.0
aniso8601==8.0.0
asn1crypto==0.24.0
attrs==19.1.0
Authlib==0.11
bcrypt==3.1.7
boto==2.49.0
boto3==1.9.234
botocore==1.12.234
bson==0.5.8
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
colormath==3.0.0
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
docutils==0.15.2
Flask==1.1.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-Testing==0.7.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.2
kiwisolver==1.1.0
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.3
numpy==1.17.2
oauth2client==4.1.3
Pillow==6.1.0
pixelate==0.4
pprint==0.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pymongo==3.9.0
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
schedule==0.6.0
scipy==1.3.1
six==1.12.0
SQLAlchemy==1.3.8
uritemplate==3.0.0
urllib3==1.25.5
Werkzeug==0.15.6
